Changelog on GitHub

+++ src/main/java/com/qa/ims/persistence/dao/OrderDAO.java
@@ -7,9 +7,8 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
-import java.util.Date;
+import java.sql.Date;
 import java.util.List;
-
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -18,6 +17,8 @@ import com.qa.ims.utils.DBUtils;
 
 public class OrderDAO implements Dao<Order>{
 
+	private OrderLineDAO orderLineDAO = new OrderLineDAO();
+	
 	public static final Logger LOGGER = LogManager.getLogger();
 	
 	@Override
@@ -47,7 +48,7 @@ public class OrderDAO implements Dao<Order>{
 	public Order readLatest() {
 		try (Connection connection = DBUtils.getInstance().getConnection();
 			 Statement statement = connection.createStatement();
-			 ResultSet resultSet = statement.executeQuery("SELECT * FROM orders ORDER BY order_id DESC LIMIT 1");) {
+			 ResultSet resultSet = statement.executeQuery("SELECT * FROM orders ORDER BY order_id DESC LIMIT 1 ");) {
 			
 			resultSet.next();
 			return modelFromResultSet(resultSet);
@@ -69,7 +70,7 @@ public class OrderDAO implements Dao<Order>{
 				Statement statement = connection.createStatement();) {
 			
 			statement.executeUpdate("INSERT INTO orders(customer_id, date_ordered, total_cost) values("
-					+ order.getCustomerID()+ "," + order.getDateOrdered()+ "," + order.getTotalCost() +");");
+					+ order.getCustomerID()+ ",'" + order.getDateOrdered()+ "'," + order.getTotalCost() +");");
 			
 			return readLatest();
 			
@@ -107,8 +108,8 @@ public class OrderDAO implements Dao<Order>{
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				Statement statement = connection.createStatement();) {
 			
-			statement.executeUpdate("update orders set customer_id = " + order.getCustomerID() + ", date_ordered = "
-					+ order.getDateOrdered() + ", total_cost = " + order.getTotalCost());
+			statement.executeUpdate("update orders set customer_id = " + order.getCustomerID() + ", date_ordered = '"
+					+ order.getDateOrdered() + "', total_cost = " + order.getTotalCost());
 			
 			return readOrder(order.getOrderID());
 			
@@ -124,10 +125,35 @@ public class OrderDAO implements Dao<Order>{
 	@Override
 	public int delete(long id) {
 		
+		if(orderLineDAO.deleteOrder(id) == 1) {
+			
+			try (Connection connection = DBUtils.getInstance().getConnection();
+					Statement statement = connection.createStatement();) {
+				
+				return statement.executeUpdate("delete from orders where order_id = " + id);
+				
+			} catch (Exception e) {
+				
+				LOGGER.debug(e);
+				LOGGER.error(e.getMessage());
+			
+			}
+			
+		}else {
+			
+			return -1;
+		}
+		
+		return 0;
+		
+	}
+	
+	public int deleteCustomer(long id) {
+		
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				Statement statement = connection.createStatement();) {
 			
-			return statement.executeUpdate("delete from orders where order_id = " + id);
+			return statement.executeUpdate("delete from orders where customer_id = " + id + ";");
 			
 		} catch (Exception e) {
 			
@@ -145,19 +171,24 @@ public class OrderDAO implements Dao<Order>{
 		Long orderID = resultSet.getLong("order_id");
 		Long customerID = resultSet.getLong("customer_id");
 		Date dateOrdered = resultSet.getDate("date_ordered");
-		double totalCost = resultSet.getDouble("total_cost");
+		Double totalCost = resultSet.getDouble("total_cost");
 		
 		return new Order(orderID, customerID, dateOrdered, totalCost);
 	}
 	
-	public Order updateTotalPriceCreate(double total) {
+	public Order updateTotalPriceCreate(Long orderID) {
 		
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				Statement statement = connection.createStatement();) {
+				Statement statement = connection.createStatement();
+				ResultSet resultSet = statement.executeQuery("SELECT SUM(product_quantity*price) AS orderline_total FROM orderline JOIN products on "
+				 		+ "orderline.product_id = products.product_id WHERE order_id = " + orderID);) {
 			
-			statement.executeUpdate("update orders set total_cost = " + total + "WHERE order_id = " + readLatest().getOrderID());
+			resultSet.next();
+			Double total = resultSet.getDouble("orderline_total");
 			
-			return readOrder(readLatest().getOrderID());
+			statement.executeUpdate("update orders set total_cost = " + total + "WHERE order_id = " + orderID);
+			
+			return readOrder(orderID);
 			
 		} catch (Exception e) {
 			
@@ -169,14 +200,20 @@ public class OrderDAO implements Dao<Order>{
 		
 	}
 	
-	public Order allOrdersByCust(Long id) {
+	public List<Order> allOrdersByCust(Long id) {
 		
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				Statement statement = connection.createStatement();
 				ResultSet resultSet = statement.executeQuery("SELECT * FROM orders where customer_id = " + id);) {
 			
-			resultSet.next();
-			return modelFromResultSet(resultSet);
+			List<Order> orders = new ArrayList<>();
+			
+			while(resultSet.next()) {
+				
+				orders.add(modelFromResultSet(resultSet));
+			}
+			
+			return orders;
 			
 		} catch (Exception e) {
 			
@@ -184,7 +221,16 @@ public class OrderDAO implements Dao<Order>{
 			LOGGER.error(e.getMessage());
 			
 		}
-		return null;
+		return new ArrayList<>();
 	}
-
+	
+	public List<Order> readOrdersByCustomer(Long custID){
+		
+		List<Order> orders = allOrdersByCust(custID);
+		for (Order order : orders) {
+			LOGGER.info(order.toString());
+		
+		}
+		return orders;
+	}
 }
